requirements.txt
numpy>=1.20.0,<1.24.0  
opencv-python-headless>=4.5.4.60,<4.7.0  
Pillow>=10.0.0  
scipy>=1.7.0 
pandas>=1.3.0  

tensorflow>=2.12.0  
torch>=2.0.0 
torchvision>=0.15.2  
scikit-image>=0.19.0  
imgaug>=0.4.0  

Flask>=2.3.0,<3.0.0
Flask-SQLAlchemy>=3.1.1
Flask-Login>=0.6.2
Flask-CORS>=5.0.1
requests>=2.31.0  

python-dotenv>=1.0.0  
python-dateutil>=2.8.2  
pytz>=2023.3  
PyYAML>=6.0  

SQLAlchemy>=2.0.21
PyMySQL>=1.1.1  

pandas>=2.0.0
numpy>=1.20.0,<1.24.0

opencv-python-headless>=4.5.4.60,<4.7.0
Pillow>=10.0.0
scikit-image>=0.19.0

tensorflow>=2.12.0
torch>=2.0.0
torchvision>=0.15.2

ipython>=8.0.0
jedi>=0.19.2


protobuf<4.0.0,>=3.20.0  
